move test to test_fsdp
